move authentication logic to concern
